empty-trash.sh - added check for /usr/bin/trash-empty

+++ empty-trash.sh
@@ -2,12 +2,12 @@
 ###############################################################################
 # Script Name  : empty-trash.sh
 # Description  : Empty the local trash directory
-# Dependencies : tree 'sudo apt install tree'
+# Dependencies : trash-cli tree 'sudo apt install trash-cli tree'
 # Arguments    : none
 # Author       : Copyright © 2023, Richard B. Romig, Mosfanet
 # Email        : rick.romig@gmail.com | rick.romig@mymetronet.com
 # Created      : 21 Nov 2023
-# Last updated : 01 Sep 2024 Version 2.3.24245
+# Last updated : 02 Sep 2024 Version 2.5.24246
 # Comments     : Run as a user cron job.
 #              : Trash directory does not exist until a file is moved to the trash.
 #              : Tested with Debian 11/12, LMDE 6, Mint 21.x, Mint 22, MX Linux 23.1, BunsenLabs 11.
@@ -37,15 +37,19 @@ readonly log_file="trash.log"
 empty_trash() {
 	if [[ $(find "$trash_dir"/info -type f | wc -l) -gt 0 ]]; then
 		printf "\nTrash contents:\n---------------\n"
-		if exist tree; then
+		if [[ -x /usr/bin/tree ]]; then
 			tree -a "$trash_dir"
 		else
 			find "$trash_dir"/files -type f
 		fi
-		find "$trash_dir"/files -mindepth 1 -type f,d -exec rm -rf {} +
-		rm -f "$trash_dir"/info/*
-		printf "\nTrash emptied.\n"
-		if exist tree; then
+		if [[ -x usr/bin/trash-empty ]]; then
+			/usr/bin/trash-empty
+		else
+			find "$trash_dir"/files -mindepth 1 -type f,d -exec rm -rf {} +
+			rm -f "$trash_dir"/info/*
+			printf "\nTrash emptied.\n"
+		fi
+		if [[ -x /usr/bin/tree ]]; then
 			tree -a "$trash_dir"
 		else
 			find "$trash_dir"/files -type f
